requirements.txt
gymnasium>=1.0.0
numpy>=1.24.0
torch>=2.0.0
stable-baselines3>=2.0.0
matplotlib>=3.7.0
psutil>=5.9.0
scipy>=1.10.0
pyyaml>=6.0
streamlit>=1.30.0
mlflow>=2.10.0
pandas>=2.0.0
